# swagger.json
{
   "openapi": "3.0.3",
    "info": {
      "version": "1.0.0", 
      "title": "Hello Cloud Run CRUD",
      "description": "My User Project Application API",
      "contact":{
            "name": "OpenWeatherMap API",
            "url": "https://openweathermap.org/api",
            "email": "some_email@gmail.com"
          },
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }

    },
    "host": "localhost:3333",
    "basePath": "/api/v1/",
    "tags": [
      {
        "name": "Users",
        "description": "API for users in the system"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "servers": [
      {
        "url":  "http://localhost:3333/api/v1/",
        "description": "Development server"
      },
      {
        "url": "https://api-service-enduro-ulrixj4liq-as.a.run.app/api/v1/",
        "description": "Production server"
      }
    ],
    "paths": {
      "/todos": {
        "get": {
          "tags": [
            "Todos"
          ],
          "description": "get Todos in system",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "get todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        } 
      },
      "/todo/{id}": {
        "get": {
          "tags": [
            "Todos"
          ],
          "description": "get Todos in system",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of pet to return",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "get todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        } ,
        "put": {
          "operationId": "updateTodo",
          "produces": [
            "application/json"
          ],
          "tags": [
            "Todos"
          ],
          "description": "updated new Todo in system",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of pet to return",
              "required": true
            }
          ],
          "requestBody": {
            "required": true,
            "accept": "application/json",
            "content": {
              " application/json" : {
                "schema": {
                  "$ref": "#/definitions/Todo"
                }
              },
              " application/xml" : {
                "schema": {
                  "$ref": "#/definitions/Todo"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "todos successfully updated",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            },
            "400": {
              "description": "bad request todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        },
        "delete": {
          "operationId": "deleteTodo",
          "produces": [
            "application/json"
          ],
          "tags": [
            "Todos"
          ],
          "description": "delete new Todo in system",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of pet to return",
              "required": true
            }
          ],
          
          "responses": {
            "200": {
              "description": "todos successfully deleted",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            },
            "400": {
              "description": "bad request todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        }  
      },
      "/todo/findbystatus": {
        "get": {
          "tags": [
            "Todos"
          ],
          "description": "get Todos in system",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "status",
              "in": "query",
              "description":" Status values",
          "required": false,
          "explode": true,
          "schema":{
            "type": "boolean",
            "default": true,
            "enum":[ true , false]
          }
            }
          ],
          "responses": {
            "200": {
              "description": "get todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        } 
      },
      "/todo": {
        "post": {
          "operationId": "addTodo",
          "produces": [
            "application/json"
          ],
          "tags": [
            "Todos"
          ],
          "description": "create new Todo in system",
          "requestBody": {
            "required": true,
            "accept": "application/json",
            "content": {
              " application/json" : {
                "schema": {
                  "$ref": "#/definitions/Todo"
                }
              },
              " application/xml" : {
                "schema": {
                  "$ref": "#/definitions/Todo"
                }
              }
            }
          },
     
          "responses": {
            "200": {
              "description": "todos successfully created",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            },
            "400": {
              "description": "bad request todos",
              "schema": {
                "$ref": "#/definitions/Todo"
              }
            }
          }
        } 
      },
        "/divisions": {
            "get": {
                "tags": [
                  "Division"
                ],
                "summary": "Get all Division in system",
                "responses": {
                  "200": {
                    "description": "OK",
                    "schema": {
                      "$ref": "#/definitions/Divisions"
                    }
                  }
                }
              }
          },                
        "/tournaments": {
            "get": {
                "tags": [
                  "Tournament"
                ],
                "summary": "Get all Tournament in system",
                "responses": {
                  "200": {
                    "description": "OK",
                    "schema": {
                      "$ref": "#/definitions/Tournament"
                    }
                  }
                }
              }
          }
      },
     
      "definitions": {
       
        "User": {
          "required": [
            "name",
            "_id",
            "companies"
          ],
          "properties": {
            "_id": {
              "type": "integer",
              "uniqueItems": true
            },
            "isPublic": {
              "type": "boolean"
            },
            "name": {
              "type": "string"
            },
            "books": {
              "type": "array",
              "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    }
                  }
              }
                  
            },
            "companies": {
              "type": "array",
              "items": {
                  "type": "string"      
              }
                  
            }
           
          }
        },
        "Player": {
          "required": [
            "name",
            "companies"
          ],
          "properties": {
            "isPublic": {
              "type": "boolean"
            },
            "name": {
              "type": "string"
            },
            "books": {
              "type": "array",
              "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    }
                  }
              }
                  
            },
            "companies": {
              "type": "array",
              "items": {
                  "type": "string"      
              }
                  
            }
           
          }
        },
        "Division": {
          "type": "array",
          "$ref": "#/definitions/User"
        },
        "Todo": {
          "type": "object",
          "required": ["text"],
          "properties": {
            "text": {
              "type": "string"
            },
            "completed": 
            {
              "type": "boolean"
            }
        }
      }
       
      }
  }
 
  